Telegram_Chatbot_Lunch_Selection/requirements.txt
certifi==2020.6.20
numpy==1.18.5
pip==20.0.2
python-telegram-bot==12.7
requests==2.23.0
setuptools==47.1.1
urllib3==1.25.9

